+++ pkg/proxy/utils.go
@@ -1,6 +1,8 @@
 package proxy
 
 import (
+	"bytes"
+	"encoding/gob"
 	"fmt"
 	"net/url"
 	"os"
@@ -9,6 +11,8 @@ import (
 	"github.com/saucelabs/customerror"
 	"github.com/saucelabs/forwarder/internal/validation"
 )
+
+var ErrFailedToCopyOptions = customerror.NewFailedToError("deepCopy options", "", nil)
 
 // Loads, validate credential from env var, and set URI's user.
 func loadCredentialFromEnvVar(envVar string, uri *url.URL) error {
@@ -24,6 +28,22 @@ func loadCredentialFromEnvVar(envVar string, uri *url.URL) error {
 		cred := strings.Split(credentialFromEnvVar, ":")
 
 		uri.User = url.UserPassword(cred[0], cred[1])
+	}
+
+	return nil
+}
+
+// Copy from `source` to `target`.
+//
+// Basic deep copy implementation.
+func deepCopy(source, target interface{}) error {
+	buf := &bytes.Buffer{}
+	if err := gob.NewEncoder(buf).Encode(source); err != nil {
+		return customerror.Wrap(ErrFailedToCopyOptions, err)
+	}
+
+	if err := gob.NewDecoder(buf).Decode(target); err != nil {
+		return customerror.Wrap(ErrFailedToCopyOptions, err)
 	}
 
 	return nil